requirements.txt
pika==1.1.0
requests==2.22.0
Pillow==6.1.0
cx_Oracle==7.2.2
Flask==1.1.1
flask_cors==3.0.8
flask_restful==0.3.7
protobuf==3.9.1
imutils==0.5.3
jsonschema==3.0.2
mysql_connector_repackaged==0.3.1
numpy==1.17.1
pdf2image==1.7.0
pytesseract==0.3.0
scipy==1.3.1
gevent_socketio==0.3.6
